Move authcheck view into core, add login view
