+++ web_ui/views.py
@@ -11,14 +11,3 @@ class MainView(View):
 
     def get(self, request):
         return render(request, "web_ui/mainpage.html")
-
-
-class AuthCheckView(View):
-    """ View checks the current state of the user's session """
-    def get(self, request):
-        user_dict = {
-            "isAuthenticated": request.user.is_authenticated,
-            "isAdmin": request.user.is_staff,
-            "userName": request.user.username,
-        }
-        return HttpResponse(dumps(user_dict))
